// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  betterAuth: typeof betterAuth;
  discounts: typeof discounts;
  kitchen: typeof kitchen;
  menuItems: typeof menuItems;
  orders: typeof orders;
  transactions: typeof transactions;
  users: typeof users;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;